Bug 1704042 part 2 - Use GetLengthProperty overload returning uint64_t for function calls. r=lth
These clamp to or check for ARGS_LENGTH_MAX so there's no change in behavior.

Depends on D111381

Differential Revision: https://phabricator.services.mozilla.com/D111382

+++ js/src/debugger/Object.cpp
@@ -981,11 +981,11 @@ bool DebuggerObject::CallData::applyMethod() {
 
     RootedObject argsobj(cx, &args[1].toObject());
 
-    unsigned argc = 0;
+    uint64_t argc = 0;
     if (!GetLengthProperty(cx, argsobj, &argc)) {
       return false;
     }
-    argc = unsigned(std::min(argc, ARGS_LENGTH_MAX));
+    argc = std::min(argc, uint64_t(ARGS_LENGTH_MAX));
 
     if (!nargs.growBy(argc) || !GetElements(cx, argsobj, argc, nargs.begin())) {
       return false;
